For EXP-2012: Move tests for http-client into their own file
