+++ components/nimbus/src/client/mod.rs
@@ -3,7 +3,7 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 mod fs_client;
-mod http_client;
+pub mod http_client;
 mod null_client;
 use crate::error::{NimbusError, Result};
 use crate::Experiment;
