+++ components/nimbus/src/tests/mod.rs
@@ -10,3 +10,7 @@ mod test_lib_bw_compat;
 mod test_lib_schema_deserialization;
 mod test_persistence;
 mod test_versioning;
+
+mod client {
+    mod test_http_client;
+}
